chore: remove all related to nx for simpler project setup
