--- client/nx.json
@@ -1,57 +1,0 @@
-{
-	"$schema": "./node_modules/nx/schemas/nx-schema.json",
-	"npmScope": "project-portal",
-	"affected": {
-		"defaultBase": "main"
-	},
-	"tasksRunnerOptions": {
-		"default": {
-			"runner": "@nrwl/nx-cloud",
-			"options": {
-				"cacheableOperations": ["build", "lint", "test", "e2e"],
-				"accessToken": "MjI2ZTY0YTktOGUzNi00MDZiLWI4ZTctZTNkMzNkYjE4YTM5fHJlYWQtd3JpdGU="
-			}
-		}
-	},
-	"targetDefaults": {
-		"test": {
-			"inputs": ["default", "^production", "{workspaceRoot}/jest.preset.js"]
-		},
-		"build": {
-			"dependsOn": ["^build"],
-			"inputs": ["production", "^production"]
-		}
-	},
-	"namedInputs": {
-		"default": ["{projectRoot}/**/*", "sharedGlobals"],
-		"production": [
-			"default",
-			"!{projectRoot}/**/?(*.)+(spec|test).[jt]s?(x)?(.snap)",
-			"!{projectRoot}/tsconfig.spec.json",
-			"!{projectRoot}/jest.config.[jt]s"
-		],
-		"sharedGlobals": []
-	},
-	"workspaceLayout": {
-		"appsDir": "packages",
-		"libsDir": "packages"
-	},
-	"generators": {
-		"@nrwl/react": {
-			"application": {
-				"style": "styled-components",
-				"linter": "eslint",
-				"babel": true
-			},
-			"component": {
-				"style": "styled-components"
-			},
-			"library": {
-				"style": "styled-components",
-				"linter": "eslint"
-			}
-		},
-		"targetDefaults": {}
-	},
-	"defaultProject": "portal-client"
-}
