+++ client/package.json
@@ -3,26 +3,15 @@
   "version": "0.0.0",
   "license": "MIT",
   "scripts": {
-    "dev": "nx serve",
-    "start": "nx serve",
-    "build": "nx build",
-    "test": "nx test",
-    "lint": "nx run-many --all --target=lint",
-    "update:fusion": "ncu -i -f /fusion/",
-    "update:nrwl": "ncu -i -f /nrwl/"
+    "dev": "npx lerna run dev",
+    "start": "npx lerna run serve",
+    "build": "npx lerna run build",
+    "test": "lerna test",
+    "lint": "lerna run-many --all --target=lint",
+    "update:fusion": "ncu -i -f /fusion/"
   },
   "private": true,
   "devDependencies": {
-    "@nrwl/cli": "15.8.1",
-    "@nrwl/devkit": "15.8.1",
-    "@nrwl/eslint-plugin-nx": "15.8.1",
-    "@nrwl/jest": "15.8.1",
-    "@nrwl/js": "15.8.1",
-    "@nrwl/node": "15.8.1",
-    "@nrwl/nx-cloud": "15.1.1",
-    "@nrwl/react": "15.8.1",
-    "@nrwl/tao": "15.8.1",
-    "@nrwl/workspace": "15.8.1",
     "@testing-library/react": "^13.4.0",
     "@types/jest": "29.4.0",
     "@types/luxon": "^3.2.0",
@@ -42,8 +31,6 @@
     "jest": "29.4.3",
     "jest-environment-jsdom": "29.4.3",
     "lodash": "^4.17.21",
-    "nx": "15.8.1",
-    "nx-plugin-vite": "^2.2.1",
     "prettier": "^2.6.2",
     "ts-jest": "29.0.5",
     "tslib": "^2.4.0",
