+++ client/packages/notification/package.json
@@ -1,0 +1,8 @@
+{
+  "name": "@equinor/notification",
+  "version": "0.0.0",
+  "license": "MIT",
+  "scripts": {
+    "build": "tsc"
+  }
+}
