--- client/packages/notification/project.json
@@ -1,30 +1,0 @@
-{
-	"name": "notification",
-	"$schema": "../../node_modules/nx/schemas/project-schema.json",
-	"sourceRoot": "packages/notification",
-	"projectType": "library",
-	"tags": [],
-	"targets": {
-		"lint": {
-			"executor": "@nrwl/linter:eslint",
-			"outputs": [
-				"{options.outputFile}"
-			],
-			"options": {
-				"lintFilePatterns": [
-					"packages/notification/**/*.{ts,tsx,js,jsx}"
-				]
-			}
-		},
-		"test": {
-			"executor": "@nrwl/jest:jest",
-			"outputs": [
-				"{workspaceRoot}/coverage/packages/notification"
-			],
-			"options": {
-				"jestConfig": "packages/notification/jest.config.ts",
-				"passWithNoTests": true
-			}
-		}
-	}
-}
