+++ client/packages/notification/tsconfig.json
@@ -1,5 +1,5 @@
 {
-    "extends": "../../tsconfig.base.json",
+    "extends": "../../tsconfig.lib.json",
     "compilerOptions": {
         "jsx": "react-jsx",
         "allowJs": true,
@@ -12,12 +12,7 @@
         "noImplicitReturns": true,
         "resolveJsonModule": true,
         "noFallthroughCasesInSwitch": true,
-        "outDir": "../../dist/out-tsc",
-        "types": [
-            "node"
-        ]
     },
-    "files": [],
     "include": [
         "**/*.js",
         "**/*.jsx",
