--- client/packages/portal-client/project.json
@@ -1,43 +1,0 @@
-{
-	"name": "portal-client",
-	"$schema": "../../node_modules/nx/schemas/project-schema.json",
-	"projectType": "application",
-	"sourceRoot": "packages/portal-client/src",
-	"tags": [],
-	"targets": {
-		"serve": {
-			"executor": "nx-plugin-vite:serve",
-			"options": {
-				"configFile": "packages/portal-client/vite.config.ts",
-				"port": 3000,
-				"host": false,
-				"https": false
-			}
-		},
-		"preview": {
-			"executor": "nx-plugin-vite:preview",
-			"options": {
-				"configFile": "packages/portal-client/vite.config.ts"
-			}
-		},
-		"build": {
-			"executor": "nx-plugin-vite:build",
-			"options": {
-				"outDir": "../../dist/packages/portal-client",
-				"configFile": "packages/portal-client/vite.config.ts",
-				"watch": false,
-				"write": true,
-				"emitAtRootLevel": false,
-				"manifest": true
-			}
-		},
-		"test": {
-			"executor": "@nrwl/jest:jest",
-			"outputs": ["{workspaceRoot}/coverage/packages/portal-client"],
-			"options": {
-				"jestConfig": "packages/portal-client/jest.config.ts",
-				"passWithNoTests": true
-			}
-		}
-	}
-}
