--- client/packages/portal-core/.babelrc
@@ -1,12 +1,0 @@
-{
-  "presets": [
-    [
-      "@nrwl/react/babel",
-      {
-        "runtime": "automatic",
-        "useBuiltIns": "usage"
-      }
-    ]
-  ],
-  "plugins": [["styled-components", { "pure": true, "ssr": true }]]
-}
