+++ client/packages/portal-core/package.json
@@ -1,0 +1,8 @@
+{
+  "name": "@equinor/portal-core",
+  "version": "0.0.0",
+  "license": "MIT",
+  "scripts": {
+    "build": "tsc"
+  }
+}
