--- client/packages/portal-core/project.json
@@ -1,24 +1,0 @@
-{
-	"name": "portal-core",
-	"$schema": "../../node_modules/nx/schemas/project-schema.json",
-	"sourceRoot": "packages/portal-core/src",
-	"projectType": "library",
-	"tags": [],
-	"targets": {
-		"lint": {
-			"executor": "@nrwl/linter:eslint",
-			"outputs": ["{options.outputFile}"],
-			"options": {
-				"lintFilePatterns": ["packages/portal-core/**/*.{ts,tsx,js,jsx}"]
-			}
-		},
-		"test": {
-			"executor": "@nrwl/jest:jest",
-			"outputs": ["{workspaceRoot}/coverage/packages/portal-core"],
-			"options": {
-				"jestConfig": "packages/portal-core/jest.config.ts",
-				"passWithNoTests": true
-			}
-		}
-	}
-}
