+++ client/packages/portal-pages/package.json
@@ -1,0 +1,8 @@
+{
+  "name": "@equinor/portal-pages",
+  "version": "0.0.0",
+  "license": "MIT",
+  "scripts": {
+    "build": "tsc"
+  }
+}
