+++ client/packages/portal-ui/package.json
@@ -1,0 +1,8 @@
+{
+  "name": "@equinor/portal-ui",
+  "version": "0.0.0",
+  "license": "MIT",
+  "scripts": {
+    "build": "tsc"
+  }
+}
