--- client/packages/portal-ui/project.json
@@ -1,24 +1,0 @@
-{
-	"name": "portal-ui",
-	"$schema": "../../node_modules/nx/schemas/project-schema.json",
-	"sourceRoot": "packages/portal-ui/src",
-	"projectType": "library",
-	"tags": [],
-	"targets": {
-		"lint": {
-			"executor": "@nrwl/linter:eslint",
-			"outputs": ["{options.outputFile}"],
-			"options": {
-				"lintFilePatterns": ["packages/portal-ui/**/*.{ts,tsx,js,jsx}"]
-			}
-		},
-		"test": {
-			"executor": "@nrwl/jest:jest",
-			"outputs": ["{workspaceRoot}/coverage/packages/portal-ui"],
-			"options": {
-				"jestConfig": "packages/portal-ui/jest.config.ts",
-				"passWithNoTests": true
-			}
-		}
-	}
-}
