+++ client/packages/portal-utils/package.json
@@ -1,0 +1,8 @@
+{
+  "name": "@equinor/portal-utils",
+  "version": "0.0.0",
+  "license": "MIT",
+  "scripts": {
+    "build": "tsc"
+  }
+}
