--- client/packages/portal-utils/project.json
@@ -1,24 +1,0 @@
-{
-	"name": "portal-utils",
-	"$schema": "../../node_modules/nx/schemas/project-schema.json",
-	"sourceRoot": "packages/portal-utils/src",
-	"projectType": "library",
-	"tags": [],
-	"targets": {
-		"lint": {
-			"executor": "@nrwl/linter:eslint",
-			"outputs": ["{options.outputFile}"],
-			"options": {
-				"lintFilePatterns": ["packages/portal-utils/**/*.{ts,tsx,js,jsx}"]
-			}
-		},
-		"test": {
-			"executor": "@nrwl/jest:jest",
-			"outputs": ["{workspaceRoot}/coverage/packages/portal-utils"],
-			"options": {
-				"jestConfig": "packages/portal-utils/jest.config.ts",
-				"passWithNoTests": true
-			}
-		}
-	}
-}
