--- client/packages/service-message/.babelrc
@@ -1,20 +1,0 @@
-{
-  "presets": [
-    [
-      "@nrwl/react/babel",
-      {
-        "runtime": "automatic",
-        "useBuiltIns": "usage"
-      }
-    ]
-  ],
-  "plugins": [
-    [
-      "styled-components",
-      {
-        "pure": true,
-        "ssr": true
-      }
-    ]
-  ]
-}
