+++ client/packages/service-message/components/AppChips.tsx
@@ -1,10 +1,13 @@
-import { Chip } from "@equinor/eds-core-react"
-import { FC } from "react"
-import { AppReference } from "../service-message"
+import { Chip } from '@equinor/eds-core-react';
+import { FC } from 'react';
+import { AppReference } from '../types/types';
 
 export const AppChips: FC<{ apps: AppReference[] }> = ({ apps }) => {
-    return <div>
-        {apps?.map(app => <Chip key={app.key}>{app.name}</Chip>)}
-    </div >
-}
-
+	return (
+		<div>
+			{apps?.map((app) => (
+				<Chip key={app.key}>{app.name}</Chip>
+			))}
+		</div>
+	);
+};
