+++ client/packages/service-message/package.json
@@ -1,0 +1,8 @@
+{
+  "name": "@equinor/service-message",
+  "version": "0.0.0",
+  "license": "MIT",
+  "scripts": {
+    "build": "tsc"
+  }
+}
